move bcolors class to be reusable
