+++ controller_manager/controller_manager/controller_manager_services.py
@@ -29,6 +29,19 @@ import rclpy
 import yaml
 
 
+# from https://stackoverflow.com/a/287944
+class bcolors:
+    HEADER = "\033[95m"
+    OKBLUE = "\033[94m"
+    OKCYAN = "\033[96m"
+    OKGREEN = "\033[92m"
+    WARNING = "\033[93m"
+    FAIL = "\033[91m"
+    ENDC = "\033[0m"
+    BOLD = "\033[1m"
+    UNDERLINE = "\033[4m"
+
+
 def service_caller(node, service_name, service_type, request, service_timeout=10.0):
     cli = node.create_client(service_type, service_name)
